syntax fix

+++ python_introduction/finance_calculator.py
@@ -1,11 +1,11 @@
 #personal finance calculator
 #User to input their monthly income and expenses
-monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 #Calculate the monthly savings
 monthly_savings = monthly_income - total_monthly_expenses
 #Calculate Projected Annual Savings
-projected_annual_savings = int((monthly_savings * 12) + (monthly_savings *12 * 0.05))  # Assuming a 5% increase in savings each year
+projected_annual_savings = (monthly_savings * 12) + (monthly_savings *12 * 0.05)  # Assuming a 5% increase in savings each year
 #Output the results
 print(f"Your monthly savings are ${monthly_savings}.")
 print(f"Projected savings after one year, with interest, is: ${projected_annual_savings}.")
